% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Tests Knapsack using Greedy Heuristic method}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{The first parameter i.e. a data frame}

\item{W}{The second parameter i.e. weight constraint}
}
\description{
Tests Knapsack using Greedy Heuristic method
}
