% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Tests Knapsack using dynamic programming}
\usage{
knapsack_dynamic(df, W)
}
\arguments{
\item{df}{The first parameter i.e. a data frame}

\item{W}{The second parameter i.e. weight constraint}
}
\description{
Tests Knapsack using dynamic programming
}
