% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic_cpp.R
\name{knapsack_dynamic_cpp}
\alias{knapsack_dynamic_cpp}
\title{Dynamic knappsack with rcpp}
\usage{
knapsack_dynamic_cpp(df, W)
}
\arguments{
\item{df}{The first parameter i.e. a data frame}

\item{W}{The second parameter i.e. weight constraint}
}
\description{
Dynamic knappsack with rcpp
}
