% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Tests Knapsack using four different methods}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{The first parameter i.e. a data frame}

\item{W}{The second parameter i.e. weight constraint}
}
\description{
Tests Knapsack using four different methods
}
